Comprehensive testing script: restore correct Windows PATH manipulation
The Windows PATH manipulation should put the dll subdirectory in the
build tree in the PATH for build-tree tests for BOTH the shared and
nondynamic cases.  Due to a misunderstanding when rewriting this logic
after the 5.11.0 release, I changed this logic to forget about the
nondynamic case, and Arjen's recent Cygwin test failed in the
nondynamic case because of that.  This commit restores that correct
behaviour.