Traditional build system for installed examples: Improve linking for non-shared case
The problem is that libplplot is a mixed C and C++ library for the
nondynamic or static case (because of including C++ drivers code in
the library for those cases).  Therefore, for the traditional build
system for the installed examples, when gcc is used to link a C
example a specific link to libstdc++.so has to be provided.

The code in src/CMakeLists.txt to find that library is really ugly,
works only for g++ and no other C++ compiler, and depends on g++
internals which apparently have now changed.  (Uses colon-separation
of libraries rather than semicolon separation.) And figuring out how
to deal with that colon separation when there are drive letters is a
real pain to deal with.

So I have completely abandoned that method and instead simply use the
C++ compiler (which automatically links in libstdc++) for the
non-shared case when building C examples using the traditional build
system.  For some reason I had to change how plexit was declared in
examples/c/x08c.c to get this idea to work.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using

scripts/comprehensive_test.sh \
--do_ctest no \
--do_test_interactive no \
--do_test_build_tree no \
--do_test_install_tree no