libplplot: plInBuildTree: bugfix: For windows check both possible directory separators
Previously the logic assumed the directory separator was '\\' on
Windows, but with recent versions of MSYS this may have changed so
now for Windows platforms check both the '\\' and '/' separators.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on both Linux
and MinGW/MSYS/Wine by using the -debug option to make sure the
build-tree drivers directory is properly found in both cases, i.e.,
plInBuildTree was correctly identifying that the code was being run
from the build tree on both platforms.  Note, this test worked well
before because I had a moderately old MSYS version installed that only
uses the '\\' directory separator in this context.  So the definitive
test will be by Arjen where plInBuildTree was not correctly
identifying that the code was being run from the build tree so that
the build-directory drivers directory could not be found on his more
modern MSYS system.