Build system: Update minimum CMake version to 3.2.2 for most platforms
The platforms where the minimum version remains at 3.0.2 are Linux
and Cygwin, and that will likely continue for a couple of more years
until all modern Linux distros and Cygwin install a cmake version >= 3.2

The reason for requiring 3.2.2 for most platforms is there is a
permissions bug in file(GENERATE...) that is fixed in cmake 3.2, and a
straightforward workaround for that bug is only available for Linux
and Cygwin (e.g., not Mac OS X) where we are retaining the lower
minimum version in any case.

This commit also processes the plplot_test subdirectory before the
examples subdirectory so that the planned workaround_file_generate_bug
target which is designed to fix the permissions problem will work properly.