Build system: move to using CMP0026 NEW
This is a rather intrusive build-system change where use
of the deprecated LOCATION property for build targets is replaced by
using appropriate generator expressions.

One large complication of this change is that certain files now have
to be configured that contain generator expressions.  I have designed
the configure_file_generate function to do that task which combines
appropriate calls to configure_file and file(GENERATE...) to configure
files containing both the normal configure_file tokens but also
generator expressions.  This approach works fine for CMake 3.2, but
for versions less than that there is a file(GENERATE...) permissions
bug which has to be fixed with a GNU extension to chmod which is
apparently only available on Linux and Cygwin.  But that dovetails
neatly with our desire to maintain a CMake minimum version of 3.0.2
for likely the next two years for both those platforms so it is only
for those platforms where a permission fix is convenient (using the
GNU extension to chmod) that a permissions fix is required.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using

scripts/comprehensive_test.sh --cmake_added_options \
"-DDEFAULT_NO_DEVICES=ON -DPLD_tk=ON -DPLD_xwin=ON -DPLD_ntk=ON -DPLD_tkwin=ON -DPLD_ps=ON -DDEFAULT_NO_BINDINGS=ON -DENABLE_tcl=ON -DENABLE_itcl=ON -DENABLE_tk=ON -DENABLE_itk=ON"

The restrictions on the tests specified by --cmake_added_options are
to speed the test while not compromising the ability to comprehensively
test all Tcl and Tk targets which are the only ones where the
configure_file_generate function now has to be used.