Core C library: fix floating point issue in notcrossed function
The issue was that sometimes the calculated intersection was far
outside the line segments so the conversion of that floating-point
information to PLINT was out of range.  The returned intersection is
only used when status = 0 (i.e., there was a definite crossing within
both line segments) so the fix was only to do that conversion when
status = 0.

This change fixed the floating-point issues revealed by using
the gfortran option -ffpe-trap=invalid,zero,overflow,underflow'
for fortran examples 25 and 30, but floating-point issues still
remain for examples 21 and 33.