Fortran example 29: fix argument specification bug
The change was 0. ==> 0._plflt in two different plctime calls and this
cleared up the case where the seconds variable was being replaced by a
random floating point number in the C version of the plctime call.

This change also apparently fixed the floating-point exceptions for
Fortran example 29 although I leave it to Hazen (who drew attention
to the plctime issue) to explain why.