Build system: Fix linking issues with traditional build of installed examples
The general issues fixed in this commit are the following:

1. Distinctions between PUBLIC (for shared libraries) and PRIVATE for
(static libraries) linking were not being done in the configured *.pc
files leading to overlinking issues for the shared libraries case.
Now, most libraries (unless they are always supposed to be PUBLIC) are
assigned to PUBLIC or PRIVATE depending on whether NON_TRANSITIVE is
OFF or ON.

2. The linking configuration of the *.pc files was additionally
incomplete for some of our libraries leading to underlinking.

3. The link steps in the Makefiles for the traditional build were
being configured to an incomplete result leading to underlinking.

I dealt with these issues in general by transforming the linking
solutions developed in the CMake case (since those are better debugged)
to the pkg-config and configured Makefile case.

Along the way, I discovered a fixed the following issues in our
CMake build system:

A. libplplotada must always link to libplplot as PUBLIC.

B. pkg_config_link_flags function must allow empty items in the lists of
libraries it translates into pkg-config form.  (This was handled
by using quotes appropriately.)

C. Some libraries need to be treated as PUBLIC regardless (such as
case A above).  To handle this issue for the pkg-config case,
I added an optional 7th argument to pkg_config_file containing
a list of libraries that should be treated as always PUBLIC.

D. --static option now used for pkg-config results in the traditional
build when BUILD_SHARED_LIBS=OFF.  (Previously this only happened when
FORCE_EXTERNAL_STATIC was ON which was a mistake since that
experimental option is used for another purpose.)

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using default (i.e., everything tested) scripts/comprehensive_test.sh.

N.B. this only tested the Qt4 case on Debian stable so additional
comprehensive testing is requested for Qt5 and/or additional platforms.