Build system: Drop problematic specific Cygwin logic for the traditional build of the fortran examples.
This problematic logic was introduced in 2007 for f77 and inherited
for f95 later.  It specifies for the traditional build a versioned
library name for both libplplotf95c and libplplotf95 installed in a
non-standard location. This problematic logic may have worked back in
2007, but it definitely does not work now leading to traditional build
errors on Cygwin with a message stating "cannot find
-lplplotf95-12.0.0".

So as an experiment I have completely removed this logic and instead
treat the traditional build of fortran examples in the same standard
way as the C, C++, etc., examples.

N.B. needs testing on Cygwin to see whether removing this problematic
logic for Cygwin is the right approach.