Fix styling issue that caused a bad result for scripts/style_source.sh
The bad result was in include/pldll.h.in where for an #if preprocessor
directive uncrustify removed a misaligned continuation "\" character
rather than aligning it on the end of a second line of a whole bunch
of continued lines. The solution was to align the "\" by hand, and
after that change uncrustify no longer removed the continuation
character.

N.B. This is the first time in years that uncrustify has introduced a
code error, but the episode is useful because it should strongly
motivate us to always carefully check what uncrustify does (typically
with "git diff --ignore-all-space") before committing the result.