Core library: Add needed level check for calls to plP_state
Put a

if ( plsc->level > 0 )

level check on all calls to

plP_state( PLSTATE_FILL );

in src/plsdef.c. Without these level checks, then the combination
of

plschr( 0.0, 1.0);
plinit();

in that order leads to an invalid read as reported by valgrind and
a segfault.  See also <https://sourceforge.net/p/plplot/bugs/162/>.

Note the wxwidgets/plbuf changes associated with the last release
cycle moved a call to plP_state to the spat routine (but without the
level check that it had in its previous location).  So that was a
regression against 5.10.0.  Furthermore, the wxwidgets/plbuf changes
introduced two additional calls to plP_state within this source file
without the level check.  So this fix for those three cases solves a
regression and two new bugs introduced by the wxwidgets/plbuf changes.