Build system: exclude testing the wxwidgets device from the test_interactive target
The issue is the -np option does not yet work for the wxwidgets
device.  Therefore, including this device in the list of tests run by
the test_interactive target significantly increases the interactive
load (clicking on the mouse to display all the pages and eventually
exit the wxPLViewer GUI for each of the examples that are tested) when
running that target, and thus makes running that target much too heavy
a burden for interactive testers.