Added a clipper class to wxwidgets_dev.cpp. This differs from wxDCClipper in that it can deal with (i.e. does nothing for) a null dc and upon destruction it restores the clipping region to the previous state rather than removing all clipping regions.
Removed all the if( m_dc ) checks from calls to the DC state changing
objects in wxPLDevice. See previous commit for details of why this was a
problem and modifications to these classes to allow null pointers.

Now we have text back for wxPLplotDemo, but still some issues in
wxPLViewer - for x00 we have no data in the plot and only a single axis
label of 0.0