Build system: Work around linking issue for traditional static build
One minor bug fix to previous fix for this issue.  On some platforms,
e.g., Mac OS X, CMAKE_CXX_IMPLICIT_LINK_DIRECTORIES can contain not
only the usual short CMake-style library names but also the full path
name for a library. So implement an if(EXISTS...) check for this
possibility if find_library cannot find the library based on the
asumption that the CMAKE_CXX_IMPLICIT_LINK_DIRECTORIES element is a
short CMake-style library name.