Added an append buffer escape and a flush remaining buffer escape and an AppendBuffer function to wxPLplotstream. The use is pretty obvious in that PLESC_APPEND_BUFFER appends data to the end of the current buffer and PLESC_FLUSH_REMAINING_BUFFER flushes any buffer that hasn't been rendered. A slight restructuring of the plRemakePlot() function in plbuf.c allowed this.
The impact is a massive speed increase of wxPLViewer when using lots of
flusheses, for example the stripchart in example 17 ground to a halt in
the previous version which used PLESC_IMPORT_BUFFER and plreplot to
replot the whole buffer after each flush.

Also added an esc_old variable in plbuf_control. This is just useful for
debugging buffer coruption issues and isn't actually used by the library
(so will probably get optimised away anyway in release builds)