Build system: bug fix for plplotada library naming convention
For some reason (not yet analyzed but my guess is the problem occurs
for every Windows platform where UNIX is not true, i.e., all other
platforms than Cygwin) the naming convention for the plplotada library
on MinGW is not consistent with our other libraries so there has been
a long-standing workaround for that issue.

That workaround errors out on Cygwin because on that platform a very
different naming convention is used.  Furthermore, it appears the
naming convention for the plplotada library on Cygwin is consistent
(i.e., no naming-convention workaround of any kind is needed for that
platform).

Thus, this bug fix simply drops the naming-convention workaround
for the Cygwin case.  But for all other Windows platforms a
more fundamental fix is required than the current workaround.