Build system: FindwxWidgets.cmake: order of search for aliases of wx-config
The usual CMake philosophy is to search first for most specific forms
of names in descending order by version then generic.  So I have
followed that philosophy for the possible aliases of wx-config.  I
also included 2.9 in the range of versions from 3.0 to 2.8 because
that is also a possibility. Finally, I replaced the previous logic of
separate find_program calls by using the NAMES signature for
find_program.