Build system: Avoid redundant linking of tclmatrix and plplottcltk libraries
If ENABLE_DYNDRIVERS=OFF then all code in both the tclmatrix and
plplottcltk libraries appears in the plplot library.  So for this case
we replace redundant linking to the tclmatrix, plplottcltk, and plplot
libraries by just linking to the plplot library. In fact, for this
case we don't bother to build or install the tclmatrix and plplottcltk
libraries at all.

Note, the working hypothesis behind these changes is redundant linking
might cause memory management issues on library exit.  At least that is
our experience with a similar redundant linking problem (since fixed)
for the plplotqt library.  But regardless of that potential issue, I
also feel it is cleaner not to use redundant linking.

In addition to the above fundamental and intrusive change, we have
taken this opportunity to also do the following:

1. Rewrote bindings/tk/pkgIndex.tcl.in and
bindings/tk-x-plat/pkgIndex.tcl.in to be exactly the same as
bindings/tcl/pkgIndex.tcl.in aside from the name of the TEA-based
package being loaded.  The corresponding CMakeLists.txt files have
been changed to configure these files correctly in all three cases.
This approach works on Unix (which fixes a bug where the Pltk and
PLplotter packages would not load for the ENABLE_DYNDRIVERS=OFF case
before) and should also work on Windows, but that latter platform
remains to be tested.

2. Updated instructions in examples/tcl/README.tcldemos and
examples/tk/README.tkdemos about running the Tcl/Tk/Itcl/Itk
examples by hand (especially useful for valgrind runs).

3. I don't think it is an improvement due to dropping redundant linking
but for examples/tk/wish_standard_examples(.in) it appears examples
2 and 24 now work without memory management issues or any other kind
of run-time error and example 20 works also without nopause issues.
Therefore, I have added 2, 20, and 24 to the examples that are run
by this script.

4. Add test_pltcl_standard_examples to the dependencies of test_tcl
and test_interactive targets for the Windows case. I have made this
change because we have good memory management results for
test_pltcl_standard_examples on the Linux platform according to
valgrind, and Arjen's previous test of this target where he reported
memory management issues on Cygwin may be an artifact of his previous
inconsistent Tcl/Tk/Itcl/Itk installation on that platform.

5. Sort out visibility issues for the tclmatrix and plplottcltk code
that is included in the plplot library when ENABLE_DYNDRIVERS=OFF.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using the command

scripts/comprehensive_test.sh --cmake_added_options "-DDEFAULT_NO_DEVICES=ON -DPLD_tk=ON -DPLD_xwin=ON -DPLD_ntk=ON -DPLD_tkwin=ON -DPLD_ps=ON -DDEFAULT_NO_BINDINGS=ON -DENABLE_tcl=ON -DENABLE_itcl=ON -DENABLE_tk=ON -DENABLE_itk=ON"

I also tested the tclsh and wish examples in
examples/tcl/README.tcldemos and examples/tk/README.tkdemos using
valgrind.  For some still unknown reason (not related to redundant
linking which is now gone) there continues to be severe memory
management issue when attempting to exit wish after running the
standard examples under wish.  All other results including those from
examples 2, and 24 were valgrind clean (until the end) which is why I
now include example 2 and 24 in the test, see comment 3 above.