Build system: Fix potential inconsistencies in itcl and itk variables
With the old CMake logic it was possible in a certain corner case
(found by one of Arjen's tests when an iwidgets package was not
installed on Cytwin) for HAVE_ITCL, HAVE_ITCLDECLS_H to be set
inconsistently with ENABLE_itcl and HAVE_ITK inconsistently with
ENABLE_itk.  The result was a build error with missing symbols which I
verified also on Linux. This fix makes such inconsistencies
impossible, and therefore should make it possible to build PLplot when
iwidgets is not found.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by temporarily looking for iwidgetsxx rather than iwidgets to
emulate a missing iwidgets.  The result was bad

make plplottcltk
ldd -r bindings/tcl/libplplottcltk_Main.so

results before the fix, and good results afterwards.