Build system: fix logic bugs that were revealed by the -DPLD_tk=OFF -DPLD_tkwin=OFF case and the -DENABLE_itk=OFF case
1. Change

NOT ENABLE_DYNDRIVERS ==> tcltk_in_plplot_library

everywhere as a test whether the tclmatrix and plplottcltk code
is compiled as part of the plplot library or whether the tclmatrix
and plplottcltk libraries are actually built and linked to separately.

The tcltk_in_plplot_library variable is equivalent to

NOT ENABLE_DYNDRIVERS AND (PLD_tk OR PLD_tkwin)

and the effective addition of that AND (PLD_tk OR PLD_tkwin) is the real
fix here.

This change fixes at least an undefined symbol issue with the
plplottcltk_Main library that shows up for the -DPLD_tk=OFF
-DPLD_tkwin=OFF case, but I suspect there are other corner cases that
are fixed by this change as well.

2. Change to setting the -DUSINGDLL compile flag on library source
code so that occurs just before when the library is built to make sure
that flag is used for all source code for a given library.  This
change at minimum fixes a visibility bug in the build of the
plplottcltk_Main library that showed up when -DPLD_tk=OFF
-DPLD_tkwin=OFF, but I suspect there are other corner cases that are
fixed by this change as well.

3. Test PLD_tk rather than ENABLE_tk when deciding
whether to configure bindings/tk/pkgIndex.tcl to load the TEA-based
Pltk package.  That package is based exactly on the tk device driver
code so this fix solves a bug where CMake errors out at generation
time with the message

CMake Error at cmake/modules/plplot_functions.cmake:251 (file):
  Error evaluating generator expression:

    $<TARGET_FILE_DIR:tk>

  No target "tk"
[...]

for the corner case where PLD_tk=OFF and ENABLE_tk=ON

4. Fix several issues with whether various Tcl/Tk/Itcl/Itk tests are
configured or not.

4a. test_tk_0[1-4], test_tk_plgrid, and test_plserver_standard_examples
no longer incorrectly depend on PLD_tk.

4b. test_plserver_runAllDemos no longer incorrectly depends on PLD_tk and
ENABLE_itkX.

4c. test_wish_runAllDemos now correctly depends on PLD_tkwin and no longer
incorrectly depends on PLD_tk and ENABLE_itkX.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux using

scripts/comprehensive_test.sh --cmake_added_options "-DDEFAULT_NO_DEVICES=ON -DPLD_tk=OFF -DPLD_xwin=ON -DPLD_ntk=ON -DPLD_tkwin=OFF -DPLD_ps=ON -DDEFAULT_NO_BINDINGS=ON -DENABLE_tcl=ON -DENABLE_itcl=ON -DENABLE_tk=ON -DENABLE_itk=ON"

N.B. note the -DPLD_tk=OFF -DPLD_tkwin=OFF options in the above command.

In addition, the test_plserver_runAllDemos and test_wish_runAllDemos
test targets (which are currently excluded from comprehensive testing
because of some issues with runAllDemos.tcl) were built for the
-DENABLE_itk=OFF case proving those tests could be run (as well as
usual) independent of Itk.