Impliment use of nopause in the wxWidgets driver
Added a chek for nopause for the wxwidgets driver when using wxViewer.
This was a missing feature to now. To do this the device sends a new
transmissionEndOfPageNoPause ratehr than a transmissionEndOfPage
command. This causes the viewer to immediately display the transmitted
page rather than waiting for user input.
In addition a transmissionClose has been added which causes the viewer
to exit. if pls->nopause is set then this command will be sent just
before the device destructor is called and it will cause the viewer to
close.