Optimisation of wxPLViewer to speed up plotting
We were getting very slow rendering which proved to be related to requesting
text size from the viewer. To improve this I removed a mutex from
wxDevice::DrawTextSection as it was not needed and now do not immediately slow
the the buffer poling in wxPLViewer if it finds no data.

Signed-off-by: Phil Rosenberg <p.d.rosenberg@gmail.com>