Changed the way fonts are passed to wxPLViewer and created when received
This fixes a but when using Ubuntu with wxWidgets 3.0 and hopefully
gives a speed increase.

When the console mode part of the app created a wxFont to get a text
description to pass to wxPLViewer, this caused a crash on the above
system. Instead now we simply convert the plplot native font parameters
to those required for a wxFont and pass these instead. When the
wxPLViewer receives the parameters it uses wxTheFontList to create the
font. This caches fonts so speeds up multiple calls.