Optimisation of wxPLViewer
We now check for new commands multiple times each time we get a timer
event. The reading of the transmission data has been moved to a separate
funtion which returns false if it checked the buffer and found there was
nothing to read and returns true otherwise. When we receive a timer
event we repeatedly call this function until we get a false. On Windows
with wxWidgets 3.0 the speed up is not huge, but on CentOS with
wxWidgets 2.8 there is a ten times speed increase for example 8. The
difference is presumably to do with the different timers and event
handling between the two systems.