Removed exit() calls from wxPLViewer
There were two exit calls in wxplframe.cpp. Calling exit bypasses all
the cleanup code including freeing any system resources that have been
allocated. The exit call in the constructor of wxPlFrame has been
replaced with a throw, which is caught in wxPlViewerApp::OnInit, which
then returns false causing the application to close with an error code.
The exit call IN wxPlFrame::ReadTransmission has been replaced by a call
to wxWindow::Close() which causes the frame to close gracefully and
because it is the only top level window this causes the application to
close too.