Swig-generated bindings documentation: Make that consistent with doc/docbook/src/api.xml
This is done by running

make check_swig_documentation

which generates from doc/docbook/src/api.xml documentation for the
swig-generated bindings and warns if that generated result is
different from the source-tree result.  For this case where that
warning occurred, I diffed the two results to make sure the generated
version was OK, then copied that generated version back on top of the
source-tree version.