Build system: Bump CMake minimum version from 3.2.2 to 3.2.3
This change is motivated by what looks like an important fix in 3.2.3
(see <http://www.cmake.org/Bug/view.php?id=15546>) which works around
mingw32-make trailing backslash trouble which in turn (without the
workaround) interferes with the use of the "MinGW Makefiles" generator
for a number of different compilers.

The CMake minimum version remains at 3.0.2 for the Cygwin and Linux
platforms, i.e., they are unaffected by this change.