Add MSYS to the condition for including a .def file in the linking of the Fortran bindings on Windows.
Rationale:
If you use the MSYS shell, then CMake defines MSYS but not MINGW to identify theplatform, but the underlying compiler (gfortran) is the same. Hence to make sure that the right link information is used (in this case no .def file should be added to the link step), we need to cover the MSYS environment as well.

Tested on MingW-w64/MSYS2