epa_build: configure octave-lite
This is an epa_build of octave-4.0.0 that is configured to drop a
number of optional parts of octave.

As part of this change move the epa_build of lapack_blas from
the BUILD_THE_BUILD_TOOLS=OFF case to the BUILD_THE_BUILD_TOOLS=ON case
since the configuration of octave_lite (when
BUILD_THE_BUILD_TOOLS=OFF) requires that lapack_blas be built already.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux by

epa_building octave_lite and doing superficial run-time tests of
that result such as

octave:1> x=-pi:0.1:pi;
octave:2> plot(x,sin(x));

This test worked fine with default plotting (gnuplot), but segfaults
if a locally modified epa_build of plplot_lite is used that enables
octave and when the plotting is done with PLplot (using the
appropriate .octaverc such as the one configured in bindings/octave
for the PLplot build).

This same PLplot run-time test works fine when the octave binding of
PLplot is built against the system octave-3.6.2.  I assume the
segfault for the octave-4.0.0 PLplot plotting case is caused by some
incompatibility between the octave binding of PLplot and octave-4.0.0,
rather than some issue with the epa_build of octave-4.0.0, but further
work will be needed to confirm that assumption.