Improve comprehensive_test.sh script
The improvements are the following:

1. Make the summary of options state the options
with the same format users would need to specify them (and with quotes
wherever they might be required).

2. Make the script more robust by warning when no printenv command is
available, for that case removing corresponding stale file results, and
by checking that files exist before storing them in the tarball.

when there is no printenv command
available.