Build system: remove garbage from MINGWLIBPATH and BORLANDLIBPATH variables
These variables have been set with incorrect set command syntax
(i.e., with an extra DOCSTRING directive before the actual doc string)
from the very first.  The result is the set command is interpreted as
specifying a CMake list as the value of the ordinary uncached
MINGWLIBPATH and BORLANDLIBPATH variables.  That list for MINGWLIBPATH
is

<desired result>;CACHE;FILEPATH;DOCSTRING;"Path to MinGW import libraries"

and similarly for BORLANDLIBPATH.

These extraneous extra list values were than used as garbage HINTS to
help find wingcc libraries which did not affect the actual finding
since the <desired result> was also a HINT.  So the net effect of this
bad set syntax was the libraries for wingcc were still found with a good
HINT plus a lot of garbage HINTS, but MINGWLIBPATH and BORLANDLIBPATH
are uncached, i.e., impossible for the user to modify.

These issues have now been addressed by removing the invalid extra
DOCSTRING directives in the set commands and the result is the garbage
hints are gone, and the results for MINGWLIBPATH and BORLANDLIBPATH
are now properly cached (as discovered with a mini test project on
Linux).