Fix problem with wxFont affecting Ubuntu withwxWidgets 3.0
Creating a wxFont without a valid application in Ubuntu using wxWidgets
3,0 causes a crash. To work around this I have created a Font class
which encapsulates a wxFont and the plplot font metrics. This allows the
Font to be created and compared to other Fonts for caching puposes, but
the actual creation of a wxFont object is delayed until the wxFont is
requested. We can check for a valid wxDC bbefore we do thisavoiding the
problems.