Update of the release notes for the forthcoming 5.11.1 release
This update (which is likely the last update prior to the 5.11.1
release) includes a summary of all the improvements made during this
release cycle (based on looking at all the commit messages for this
release cycle).  This update also adds a report of comprehensive test
success for the -DOLD_WXWIDGETS=ON case.