5.11.1 preparation: Update all files that the check_all target indicates are inconsistent with other PLplot files
The check targets that are built as dependencies of the check_all
target typically generate a file in the build tree from some source
file and then compare that generated file with the corresponding file
in the source tree that was previously generated in exactly the same
way.  If that comparison fails, then the usual step is to check the
differences, and if those seem reasonable (i.e., no mistake in the
source file that was used to generate the file) then cure the issue by
copying the latest generated file in the build tree on top of the
older generated file in the source tree.  For the
bindings/swig-support/plplotcapi.i case, the comparisons are only
relevant to the one part of that file (the #defines) that is generated
so the cure in that case I edited
bindings/swig-support/plplotcapi.i to maintain consistency in the

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
both

make check_all

in the build tree and

scripts/comprehensive_test.sh --do_test_interactive no

in the source tree.