Build system: libcsironn: libplplot: implement choice between qhull/qhull_a.h and libqhull/qhull_a.h form of include for qhull headers
Because libqhull uses common names for header files (e.g.,
io.h which is also a standard Windows header) the #include must be of
the form

where <unique_name> solves the name clash between qhull headers and
other headers.  The qhull documentation recommends "qhull" for the
unique name, and most software distributions have adopted that, but
some of them (e.g., Cygwin and openSUSE) use "libqhull" instead for
the unique name.  So this commit implements an automatic choice between those
two variants where both forms are looked for, and the one that is
found is used for compilations of libcsironn and libplplot.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux by
building the csironn and plplot targets for the "qhull" case that is
appropriate to my Debian Linux distro.  N.B. Tests of this change for
distributions of software using the "libqhull" unique name convention
also need to be done.