Traditional build system: fix linking to rt library
For ENABLE_DYNDRIVERS=OFF case, wxwidgets device code is part of
libplplot, and therefore that library must be linked to the rt
library.  This linking issue has long since been solved for the CMake-based build
system case, and what the current fix does is propagate that rt
library linking fix to the configured pkg-config plplot.pc file and
therefore to the traditional build of the installed examples as well.

For some reason this linking issue for the traditional build has not
caused any problems on Debian wheezy, but it is hoped this fix will
solve a rt-related linking problem for the traditional build in the
static case that comprehensive testing by Greg Jung on the openSUSE
platform has turned up.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux using

scripts/comprehensive_test.sh --cmake_added_options "-DDEFAULT_NO_DEVICES=ON -DPLD_wxwidgets=ON -DPLD_ps=ON -DDEFAULT_NO_BINDINGS=ON -DENABLE_wxwidgets=ON -DENABLE_cxx=ON" --do_test_interactive no

which merely proves the fix does no harm in the Debian wheezy case.
However, I also checked the pkg-config results now contain the -lrt
link flag, and also that flag (as expected) is used for the
traditional build of, e.g., the C examples so the fix is working as
designed with the only question remaining concerning this fix being
whether that design will completely fix the openSUSE issue.