Added default page sizes to plplotP.h
There are two defaults. One for drivers which have real world pages
(e.g. the postscript driver) which is specified in mm and one for
drivers which use pixel pages (e.g. the Cairo or the interactive
drivers). The mm page size is A4 as this is the most commonly used
metric paper size used. The pixel size is approximately A5 when scaled
using the default 90 dpi. This was chosen as it has the same aspect
ratio as the mm size, but fits comfortably on most computer screens.