Build system and epa_build: Changes that make epa_built octave binding of PLplot work
The changes consist of the following:

1. Build system bug fix: replace all PATHS by HINTS in the find commands used to
find octave components.  This insures a consistent find result when
searching for a special version of octave that is not in one of the
standard locations.  This fix solves a segfault that otherwise occurs
when PLplot attempts to use the epa_built version of octave.

2. epa_build: Decrement octave version to 3.8.2 from 4.0.0. This may
not have been necessary to get a good result since the previous 4.0.0
segfaults for octave PLplot examples were likely due to issue 1.
Nevertheless, octave-4.0.0 is backwards-incompatible with octave-3.x.y
so the more conservative choice is to try octave-3.8.2 for now.

3. epa_build: update plplot and plplot_lite configurations to use
epa_built octave.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by building the test_noninteractive target in the build tree of
an epa_built plplot_lite that depends on epa_built octave.
There were no run-time issues at all, and the octave and C results for
the standard examples were identical.