Removed PLStream::pageset
This varible was set by c_setpage to indicate that the page parameters
had been set, however calls to this function which use 0s do not set
parameters so when this variable is set there is actually no guarentee
that the page parameters are in a usable state. It turned out that this
parameter is not used in any plplot or driver code so given its lack of
usefulness I have deleted it.