Rewrite of the wxWidgets text processing methods
The text processing routines were left unchanged in the recent driver
rewrite. The changes simplify the code, making it more readable. This
change fixes two bugs from the tracker:
subscript and superscript scales and offsets, however a multiple of
0.75 is made to the offset otherwise it doesn't match other devices
and just looks too large (it is 100% of the text height otherwise).
which copied the text into the memory buffer had been somehow delete so
wxPLViewer always saw an empty string.

In terms of restructuring. The text is dealt with in three stages, as
before, but the responsibilites of each stage are more clearly defined.

DrawText receives the full string. It sets up the dc state to be ready
to render text. It only sets up items that cannot be changed mid string.
It splits the string into lines and passes each line to DrawLine.

DrawTextLine converts the 32 bit ucs4 encoded string into a wxString. As
it does so it parses escape sequences and breaks the string at format
changes. Each of these sections is then passed to DrawTextSection.

DrawTextSection is dumb in that it has no concept of formatting. It
simply draws the text that it is tolde to, where it is told to draw it.
It has no concept of formatting such as superscript. It's only
responsibility is to select the font based on the given fci and size
and then draw the text/calculate the text size.

It should be noted that there is a commented out peice of code in
DrawText that corrects for aspect ratio of the plot based on the
assumption that plplot core code thought the plot was square. This fixes
3d text, but breaks regular 2d rotated text. The final solution of how
to deal with this needs further exploration of the core code to discover
why.