Fixed 3d text in wxWidgets driver - but with a fudge.
It seems that the rotated text code only uses device coordinates.
Because the wxWidgets driver reports the same device size (max int) in
each direction, this causes issues.
By contrast the rotated text uses mm sizes. Because we report different
device pixels per mm for x and y this worked fine.
To fix this we report identical pixels per mm in both directions then
make a correction to the rotation and shear angles in DrawText()